Refactor: 습관 이미지 모달 분리
